src/requirements.txt
nltk==3.7
pandas==1.4.3
psycopg2_binary==2.9.3
pyspark==3.3.0
requests==2.28.1
streamlit==1.10.0
tweepy==4.10.0
kafka-python==2.0.2
streamlit-autorefresh==0.0.1